gemini-med-lit-review/frontend/requirements.txt
streamlit==1.33
google-cloud-aiplatform==1.44.0 #1.48.0
langchain==0.1.15 #0.1.16
langchain-google-vertexai==0.1.2 #==1.0.1
langchain-community==0.0.32 #0.0.34
# pgvector
# psycopg2-binary
cloud-sql-python-connector[pg8000]==1.8.0
